# backend/scratch/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIHire Recruitment Platform",
    "description": "Backend API service for the AI-powered Recruitment Platform.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/proctoring/start": {
      "post": {
        "tags": [
          "AI Proctoring"
        ],
        "summary": "Start or resume monitoring session",
        "operationId": "start_session_proctoring_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartMonitoringRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/proctoring/stop": {
      "post": {
        "tags": [
          "AI Proctoring"
        ],
        "summary": "Stop and finalize monitoring session",
        "operationId": "stop_session_proctoring_stop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopMonitoringRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/proctoring/violation": {
      "post": {
        "tags": [
          "AI Proctoring"
        ],
        "summary": "Report a browser violation",
        "operationId": "report_browser_proctoring_violation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportViolationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/proctoring/frame": {
      "post": {
        "tags": [
          "AI Proctoring"
        ],
        "summary": "Submit base64 webcam frame for face/object checks",
        "operationId": "submit_webcam_frame_proctoring_frame_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportFrameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/proctoring/status": {
      "get": {
        "tags": [
          "AI Proctoring"
        ],
        "summary": "Get current monitoring status",
        "operationId": "fetch_status_proctoring_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          },
          {
            "name": "assessment_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assessment Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proctoring/history/{candidate_id}": {
      "get": {
        "tags": [
          "AI Proctoring"
        ],
        "summary": "Get candidate proctoring history for recruiter dashboard",
        "operationId": "fetch_history_proctoring_history__candidate_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportFrameRequest": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "assessment_type": {
            "type": "string",
            "title": "Assessment Type"
          },
          "frame": {
            "type": "string",
            "title": "Frame"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "assessment_type",
          "frame"
        ],
        "title": "ReportFrameRequest"
      },
      "ReportViolationRequest": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "assessment_type": {
            "type": "string",
            "title": "Assessment Type"
          },
          "violation_type": {
            "type": "string",
            "title": "Violation Type"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "assessment_type",
          "violation_type"
        ],
        "title": "ReportViolationRequest"
      },
...
      "StartMonitoringRequest": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "assessment_type": {
            "type": "string",
            "title": "Assessment Type"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "assessment_type"
        ],
        "title": "StartMonitoringRequest"
      },
      "StopMonitoringRequest": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "assessment_type": {
            "type": "string",
            "title": "Assessment Type"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "assessment_type"
        ],
        "title": "StopMonitoringRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}